+++ .gitignore
@@ -1,3 +1,4 @@
 data/
 .vs
 Leo_testing_folder/classifier_csvs/
+checkpoint*
