--- checkpoint/commits/2
@@ -1,2 +1,0 @@
-v1
-{"nextBatchWatermarkMs":0}
