--- checkpoint/metadata
@@ -1,1 +1,0 @@
-{"id":"1e46d734-dec3-4d93-a3ea-e5e9230f1251"}
