--- checkpoint/sources/0/0
@@ -1,11 +1,0 @@
-v1
-{"path":"file:///C:/Users/leois/OneDrive/Documents/SFU%20TERM1/project/SIGNIT/stream_inputs/France_24_20231205231055.txt","timestamp":1701846655892,"batchId":0}
-{"path":"file:///C:/Users/leois/OneDrive/Documents/SFU%20TERM1/project/SIGNIT/stream_inputs/Al_Jazeera_20231205231056.txt","timestamp":1701846656608,"batchId":0}
-{"path":"file:///C:/Users/leois/OneDrive/Documents/SFU%20TERM1/project/SIGNIT/stream_inputs/France_24_20231205231100.txt","timestamp":1701846660901,"batchId":0}
-{"path":"file:///C:/Users/leois/OneDrive/Documents/SFU%20TERM1/project/SIGNIT/stream_inputs/Al_Jazeera_20231205231100.txt","timestamp":1701846661179,"batchId":0}
-{"path":"file:///C:/Users/leois/OneDrive/Documents/SFU%20TERM1/project/SIGNIT/stream_inputs/France_24_20231205231104.txt","timestamp":1701846665217,"batchId":0}
-{"path":"file:///C:/Users/leois/OneDrive/Documents/SFU%20TERM1/project/SIGNIT/stream_inputs/Al_Jazeera_20231205231105.txt","timestamp":1701846665993,"batchId":0}
-{"path":"file:///C:/Users/leois/OneDrive/Documents/SFU%20TERM1/project/SIGNIT/stream_inputs/France_24_20231205231109.txt","timestamp":1701846670291,"batchId":0}
-{"path":"file:///C:/Users/leois/OneDrive/Documents/SFU%20TERM1/project/SIGNIT/stream_inputs/Al_Jazeera_20231205231110.txt","timestamp":1701846670524,"batchId":0}
-{"path":"file:///C:/Users/leois/OneDrive/Documents/SFU%20TERM1/project/SIGNIT/stream_inputs/France_24_20231205231113.txt","timestamp":1701846674511,"batchId":0}
-{"path":"file:///C:/Users/leois/OneDrive/Documents/SFU%20TERM1/project/SIGNIT/stream_inputs/Al_Jazeera_20231205231114.txt","timestamp":1701846675492,"batchId":0}
