--- checkpoint/sources/0/1
@@ -1,2 +1,0 @@
-v1
-{"path":"file:///C:/Users/leois/OneDrive/Documents/SFU%20TERM1/project/SIGNIT/stream_inputs/France_24_20231205231527.txt","timestamp":1701846927558,"batchId":1}
