--- checkpoint/sources/0/2
@@ -1,2 +1,0 @@
-v1
-{"path":"file:///C:/Users/leois/OneDrive/Documents/SFU%20TERM1/project/SIGNIT/stream_inputs/Al_Jazeera_20231205231528.txt","timestamp":1701846928694,"batchId":2}
