--- checkpoint/sources/0/3
@@ -1,3 +1,0 @@
-v1
-{"path":"file:///C:/Users/leois/OneDrive/Documents/SFU%20TERM1/project/SIGNIT/stream_inputs/France_24_20231205231726.txt","timestamp":1701847047670,"batchId":3}
-{"path":"file:///C:/Users/leois/OneDrive/Documents/SFU%20TERM1/project/SIGNIT/stream_inputs/Al_Jazeera_20231205231727.txt","timestamp":1701847047832,"batchId":3}
